.Net: feat/declarative agents (#9849)
fixes #9848
This is a draft PR, at this point it's not ready for a full review.
I'd like feedback on the location of the new extension method, and on
the fact that I had to add a reference to the agents project.
Here is the rationale for placing it here: it needs access to the
copilot agent plugins extension method and a few other types.
Here is the rationale for adding the reference to the agents project: it
needs access to the kernel agent type.

---------

Signed-off-by: Vincent Biret <vibiret@microsoft.com>
Co-authored-by: Chris <66376200+crickman@users.noreply.github.com>