.Net: Add PostgresVectorStore Memory connector. (#9324)
This PR adds a PostgresVectorStore and related classes to
Microsoft.SemanticKernel.Connectors.Postgres.

### Motivation and Context

As part of the move to having memory connectors implement the new
Microsoft.Extensions.VectorData.IVectorStore architecture (see
https://github.com/microsoft/semantic-kernel/blob/main/docs/decisions/0050-updated-vector-store-design.md),
each memory connector needs to be updated with the new architecture.
This PR tackles updating the existing
Microsoft.SemanticKernel.Connectors.Postgres package to include this
implementation. This will supercede the PostgresMemoryStore
implementation.

Some high level comments about design:
- PostgresVectorStore and PostgresVectorStoreRecordCollection get
injected with an IPostgresVectorStoreDbClient. This abstracts the
database communication and allows for unit tests to mock database
interactions.
- The PostgresVectorStoreDbClient gets passed in a NpgsqlDataSource from
the user, which is used to manage connections to the database. The
responsibility of connection pool lifecycle management is on the user.
- The IPostgresVectorStoreDbClient is designed to accept and produce the
storage model, which in this case is a Dictionary<string, object?> .
This is the intermediate type that is mapped to by the
IVectorStoreRecordMapper.
- The PostgresVectorStoreDbClient also takes a
IPostgresVectorStoreCollectionSqlBuilder, which generates SQL command
information for interacting with the database. This abstracts the SQL
queries related to each task, and allows for future expansion. This is
particularly targeted at creating a AzureDBForPostgre vector store that
will enable alternate vector implementations like
[DiskANN](https://techcommunity.microsoft.com/t5/azure-database-for-postgresql/introducing-diskann-vector-index-in-azure-database-for/ba-p/4261192),
while leveraging the same database client as the Postgres connector.
-  The integration tests for the vector store utilize Docker.Net to
bring up a pgvector/pgvector docker container, which test are run
against.

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile:

---------

Co-authored-by: Rob Emanuele <robemanuele@microsoft.com>
Co-authored-by: Dmytro Struk <13853051+dmytrostruk@users.noreply.github.com>