.Net: Amazon Connector - InnerContent Metadata Support (#10086)
### Motivation and Context

Amazon connector was not exposing the breaking glass ability to expose
the underlying component used by the Bedrock SDK for retrieving extra
metadata associated with ChatCOmpletion and TextGeneration requests.
This was creating a problem for implementers wanting to capture extra
information like: stop reason, usage, request id and more...

### Description

- Add missing InnerContent support for metadata (including Usage when
available).
- Fixes #9989 
- Fix warnings (Amazon projects) in the new Analyzer version.

---------

Co-authored-by: tanaka_733 <tanaka-takayoshi@users.noreply.github.com>