.Net: Add a request index to the streamed function call update content (#10129)
### Motivation and Context: 
Today, it's impossible to determine which request a streaming function
call update belongs to or originates from. As a result, if an AI service
interaction leads to multiple requests to an AI model - let's say one
per function call - it's impossible to assemble the function call
contents from the streamed updates. This is because the streaming
function call updates from the second request override those from the
first one.

### Description
This PR adds the `RequestIndex` to the
`StreamingFunctionCallUpdateContent` class and updates the {Azure}OpenAI
connectors to set it. Finally, it updates the
`FunctionCallContentBuilder` class to construct a composite index from
`RequestIndex` and `FunctionCallIndex` and uses it to uniquely identify
the function call content updates across requests.

Closes: https://github.com/microsoft/semantic-kernel/issues/10006