.Net: Add Aspire scaffolding for agent-hosting sample (#10702)
### Motivation, Context and Description
This PR adds scaffolding for the Aspire agent-hosting sample that will
be extended in several follow-up PRs. Ninety-nine percent of the code is
generated by the Aspire Starter App template, with a few minor changes
to add an empty agent completions controller, call it from the UI, and
display the results.

**Next steps include:**
- Deploy Azure AI Service with a chat model to be used by the agent in
the next step.
- Update the AgentCompletionsController to use a chat completion agent


**Note:** This PR targets the `feature-agents-hosting` feature branch.
**Contributes to:**
https://github.com/microsoft/semantic-kernel/issues/10149