.Net: SK integration with MEAI Abstractions (Service Selector + Contents) Phase 1 (#10651)
### Motivation and Context

- Resolves partially #10319 

This change adds a new `IChatClientSelector` abstraction to work without
breaking current `IAIServiceSelector` implementations.

This was necessary because `Kernel` now will be able to resolve
`IChatClient`s which are not `IAIService` implementations, and
`OrderedAIServiceSelector` now can select not only `AIServices` but
`ChatClients` as well by implementing the `IChatClientSelector`
interface.

All other changes are related to test accepting `IChatClient`s added to
the `Service` DI container, ensuring they work as expected together with
current `IChatCOmpletionService` and `ITextGenerationService` in the
`Kernel`.