.Net: Make AzureOpenAI service/model deployment configurable (#10786)
### Motivation, Context, and Description  

This PR reads Azure OpenAI related configuration: deployment, model, and
version from a config file instead of using hardcoded values.
Additionally, it adds a configuration property that points to an AI chat
service for the agent to use. At the moment, only the Azure OpenAI
service is supported, but other services like OpenAI will be added in
one of the next PRs.
   
It also copies some values from the host configuration to environment
variables so that the Agent API service can access them. This allows us
to keep the host/deployment configuration in one place - the host rather
than having the same config in two projects: the host and the API
service. The host uses it for deployment and provisioning of Azure
resources, while the service uses it to access the provisioned
resources.
   
Next steps include:  
- Reading agent parameters from the configuration.  
- Adding support for OpenAI chat models.
   
Note: This PR targets the `feature-agents-hosting` feature branch.    
Contributes to:
https://github.com/microsoft/semantic-kernel/issues/10149