.Net: Support for text embeddings generation model (#10829)
### Motivation, Context, and Description  

This PR adds all the required infra code to register the
{Azure}OpenAITextEmbeddingGeneration service in the agent API service DI
container so it can be used later for RAG. Additionally, it adds partial
configuration functionality for the RAG scenario which will be extended
in the future PRs.

Next steps include:  
- Register a vector store
   
Note: This PR targets the `feature-agents-hosting` feature branch.    
Contributes to:
https://github.com/microsoft/semantic-kernel/issues/10149