.Net: Apply a JSON converter for exceptions when serializing chat history (#10869)
### Motivation, Context and Description

Currently, if an exception is thrown during a parsing function call or
function invocation, the exception is stored either in
`FunctionCallContent.Exception` or `FunctionResult.Result`,
correspondingly. When chat history is serialized with either of the
content classes referencing an exception, for telemetry purposes, it
fails with the exception `System.NotSupportedException: Serialization
and deserialization of 'System.Reflection.MethodBase' instances are not
supported. Path: $.Items.Exception.InnerException.TargetSite.` This fix
updates the chat history serialization to use a JSON converter to handle
exception serialization.

Closes: https://github.com/microsoft/semantic-kernel/issues/10708