.Net: don't use a mutable static counter for tests that can run in parallel (#11004)
By default, xUnit runs tests in parallel. What we were doing so far was
having a static, mutable counter that some tests were setting to 0,
while other could use in the meantime (increment or read and assert).

I've changed it to use a static dictionary of counters, where the name
of the step is the key. So each step has it's own counter.

fixes #11001