.Net: Remove document upload functionality (#11066)
### Motivation, Context and Description
This PR removes the file upload functionality that was previously added
for testing purposes, in favor of data ingestion solutions provided by
vector stores. It also removes all UI elements related to file upload,
as well as the InMemory vector store, since there is no longer a method
to hydrate it. Additionally, it maps the data model properties of the
vector store to the auto-generated properties of the AzureAISearch
index, created by the AzureAISearch data import wizard.