.Net: Make collection optional (#11085)
### Motivation, Context, and Description  

This PR makes the `CollectionName` property optional, supporting the
basic experience where no data has been uploaded to the vector store
yet. The experience will run the agent without RAG in the absence of
data. When the data is later uploaded to a vector store collection, the
collection name will be provided in the configuration, and the agent
will run with RAG.

**Note**: This PR targets the `feature-agents-hosting` feature branch.
**Contributes** to:
https://github.com/microsoft/semantic-kernel/issues/10149