.Net: Expose underlying method from kernel function (#11378)
### Motivation, Context and Description
Currently, there is no way to access the underlying method that an
instance of `KernelFunctionFromMethod` was created from or wraps. Being
able to access this underlying method is useful when additional metadata
is needed. For example, if the method has a custom attribute, you may
need to retrieve that attribute to determine the execution path.
   
This PR adopts the same approach as M.E.AI by exposing the native method
through the `public MethodInfo? UnderlyingMethod {...}` property. This
property is null by default and is only initialized for native functions
represented by the `KernelFunctionFromMethod` class.

Closes: https://github.com/microsoft/semantic-kernel/issues/11182

---------

Co-authored-by: Roger Barreto <19890735+RogerBarreto@users.noreply.github.com>