.Net: Remove custom mappers (#11530)
We obsoleted custom mappers in #11366. However, while work on
IEmbeddingGenerator integration (#10492),I realized that for proper
embedding generation, we need to generate all embeddings for an
UpsertAsync batch in one call, rather than one-by-one (which would be
many roundtrips). This means that mapping can't happen record by record,
but must work on a batch, breaking the current design of
IVectorStoreRecordMapper.

As IVectorStoreRecordMapper is getting in the way, and custom mappers is
a feature we don't think many people are using, this PR removes support
for it (making it obsolete-with-error).