.Net: Mcp sampling sample (#11533)
### Motivation, Context and Description
This PR:
1. Adds a new sample demonstrating how to use the MCP sampling and
enabling human-in-the-loop client-side to intercept the sampling request
to decide whether to proceed with it or not.
2. Moves embedded resources from different folders to the
`ProjectResources` folder to keep all resources in one place.
3. Fixes the MCP sample that demonstrates how to use MCP tools with
ChatCompletionAgent. The sample now uses execution settings with the
enabled "retain types" option, allowing it to successfully call the
tools.

Contributes: https://github.com/microsoft/semantic-kernel/issues/11199