.Net: add 'RetainArgumentTypes' to agent arguments for ModelContextProtocolPlugin demo (#11488)
### Motivation and Context

Extremely high token consumption when running this demo:
[https://github.com/microsoft/semantic-kernel/tree/main/dotnet/samples/Demos/ModelContextProtocolPlugin](https://github.com/microsoft/semantic-kernel/tree/main/dotnet/samples/Demos/ModelContextProtocolPlugin)

When calling the ChatCompletionAgent with the default 'false' value for
RetainArgumentTypes, all arguments are converted to string values which
causes some of the tool calls to fail. The agent then iterates through
other tools, searching for a successful response. This results in 80-85k
tokens being consumed where only 11-12k should be consumed.

### Description

This PR will update the ModelContextProtocolPlugin demo to re-use the
configured PromptExecutionSettings object that includes the correct
value for `RetainArgumentTypes`.