.Net: Introduce a LINQ filter preprocessor in MEVD (#11702)
This is a refactor-only change, with no functional changes. It extracts
out all logic for identifying and processing captured variables
("parameters") out to a new FilterTranslationProcessor connectos support
type in MEVD.

* Most databases don't support parameterization; for these we simply
inline the captured variables. Connector translators see these as
constants without ever having to know about captured variables etc.
* SQL databases do support parameterization. For these, the preprocessor
transforms the captured variable (which is a complex tree construct)
into a simple QueryParameterExpression, which can easily be
pattern-matched in connector translator code.

This is preparatory work for #11673 (we need a centralized place to do
this kind of thing).