.Net: Agent Bedrock update to use abstractions (#11833)
### Motivation and Context

When running the UnitTests in a non-configured machine the tests just
halt and don't work when the mock uses the concrete types. After some
analysis this lead to the conclusion that the tests and the
implementation should be relying on the abstractions (available) instead
of the concrete types.

This change allows the Bedrock package to be compatible with the
interfaces and also simplify and make the environment light weight for
unit testing.