.Net: Add a store for textual rag (#11834)
### Motivation and Context

We should have a simple way of uploading text and doing RAG over that
text.
This branch is not up to date with the latest MEVD changes, but will be
updated after this PR with changes to make
use of the new features.

### Description

Adding a store that allows easy uploading of text or links to text and
exposing that text via the TextRagComponent

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [ ] The code builds clean without any errors or warnings
- [ ] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [ ] All unit tests pass, and I have added new tests where possible
- [ ] I didn't break anyone :smile: